refactoring utils
